refactor: move default filter logic from `callbacks` into `manio`
