+++ tests/test_manio.py
@@ -4,7 +4,7 @@ import random
 import sys
 import unittest.mock as mock
 from pathlib import Path
-from typing import Dict, List, cast
+from typing import Any, Dict, List, cast
 
 import pytest
 import yaml
@@ -12,7 +12,7 @@ import yaml
 import square.k8s as k8s
 import square.manio as manio
 from square.dtypes import (
-    Config, GroupBy, LocalManifestLists, MetaManifest, Selectors,
+    Config, FiltersKind, GroupBy, LocalManifestLists, MetaManifest, Selectors,
     SquareManifests,
 )
 from square.k8s import resource
@@ -771,7 +771,7 @@ class TestYamlManifestIO:
             assert ref == out
 
 
-class TestStripCallback:
+class TestStripManifest:
     def test_run_strip_callback_deployment(self, config, k8sconfig):
         """Filter DEPLOYMENT manifests."""
         # A valid DEPLOYMENT manifest with a few optional and irrelevant keys.
@@ -1882,3 +1882,252 @@ class TestDownloadManifests:
         assert m_get.call_count == 2
         m_get.assert_any_call(k8sconfig, res_deploy.url)
         m_get.assert_any_call(k8sconfig, res_ns.url)
+
+
+class TestManifestStripping:
+    def test_strip_manifest_generic(self, config):
+        """Create a completely fake filter set to test all options.
+
+        This test has nothing to do with real world manifests. Its only purpose
+        is to validate the algorithm that strips them.
+
+        """
+        # Define filters for this test.
+        _filters: FiltersKind = [
+            "invalid",
+            {"metadata": [
+                "creationTimestamp",
+                {"labels": ["foo"]},
+                "foo",
+            ]},
+            "status",
+            {"spec": [{"ports": ["nodePort"]}]},
+        ]
+        config.filters = {"Service": _filters}
+
+        # Demo manifest. None of its keys matches the filter and
+        # `strip` must therefore not remove anything.
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {"name": "name", "namespace": "ns"},
+            "spec": "spec",
+        }
+        assert manio.strip_manifest(config, manifest) == manifest
+        del manifest
+
+        # Demo manifest. The "labels.foo" matches the filter and must not survive.
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "name",
+                "namespace": "ns",
+                "labels": {"foo": "remove"}
+            },
+            "spec": "spec",
+        }
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "name",
+                "namespace": "ns",
+            },
+            "spec": "spec",
+        }
+        assert manio.strip_manifest(config, manifest) == expected
+        del manifest
+
+        # Valid manifest with all mandatory and *some* optional keys (
+        # "metadata.labels" and "optional-2" are present, "optional-1" is
+        # not).
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "mandatory",
+                "namespace": "ns",
+                "labels": {
+                    "foo": "remove",
+                    "bar": "keep",
+                }
+            },
+            "spec": "keep",
+            "status": "remove",
+        }
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "mandatory",
+                "namespace": "ns",
+                "labels": {
+                    "bar": "keep",
+                }
+            },
+            "spec": "keep",
+        }
+        assert manio.strip_manifest(config, manifest) == expected
+
+    def test_strip_manifest_ambigous_filters(self, config):
+        """Must cope with filters that specify the same resource multiple times."""
+        # Define filters for this test.
+        _filters: FiltersKind = [
+            {"metadata": [
+                # Specify "creationTimestamp" twice. Must behave no different
+                # than if it was specified only once.
+                "creationTimestamp",
+                "creationTimestamp",
+            ]},
+
+            # Reference "status" twice. This must remove the entire "status" field.
+            {"status": [{"foo": ["bar"]}]},
+            "status",
+        ]
+        config.filters = {"Service": _filters}
+
+        # Demo manifest. None of its keys matches the filter and
+        # `strip` must therefore not remove anything.
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {"name": "name", "namespace": "ns"},
+            "spec": "spec",
+        }
+        assert manio.strip_manifest(config, manifest) == manifest
+        del manifest
+
+        # Demo manifest. The "labels.creationTimestamp" matches the filter and
+        # must not survive.
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "name",
+                "namespace": "ns",
+                "creationTimestamp": "123",
+            },
+            "spec": "spec",
+        }
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "name",
+                "namespace": "ns",
+            },
+            "spec": "spec",
+        }
+        assert manio.strip_manifest(config, manifest) == expected
+        del manifest
+
+        # Valid manifest with a "status" field that must not survive.
+        manifest = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "mandatory",
+                "namespace": "ns",
+            },
+            "spec": "keep",
+            "status": "remove",
+        }
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {
+                "name": "mandatory",
+                "namespace": "ns",
+            },
+            "spec": "keep",
+        }
+        assert manio.strip_manifest(config, manifest) == expected
+
+    def test_strip_manifest_sub_hierarchies(self, config):
+        """Remove an entire sub-tree from the manifest."""
+        # Remove the "status" key, irrespective of whether it is a string, dict
+        # or list in the actual manifest.
+        config.filters = {"Service": ["status"]}
+
+        # Minimally valid manifest.
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {"name": "mandatory", "namespace": "ns"},
+        }
+        manifest: dict = copy.deepcopy(expected)
+
+        manifest["status"] = "string"
+        assert manio.strip_manifest(config, manifest) == expected
+
+        manifest["status"] = None
+        assert manio.strip_manifest(config, manifest) == expected
+
+        manifest["status"] = ["foo", "bar"]
+        assert manio.strip_manifest(config, manifest) == expected
+
+        manifest["status"] = {"foo", "bar"}
+        assert manio.strip_manifest(config, manifest) == expected
+
+    def test_strip_manifest_lists_simple(self, config):
+        """Filter the `NodePort` key from a list of dicts."""
+        # Filter the "nodePort" element from the port list.
+        config.filters = {"Service": [{"spec": [{"ports": ["nodePort"]}]}]}
+
+        expected = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {"name": "name", "namespace": "ns"},
+            "spec": {"type": "NodePort"},
+        }
+        manifest: dict = copy.deepcopy(expected)
+        manifest["spec"]["ports"] = [
+            {"nodePort": 1},
+            {"nodePort": 3},
+        ]
+        assert manio.strip_manifest(config, manifest) == expected
+
+    def test_strip_manifest_lists_service(self, config):
+        """Filter the `NodePort` key from a list of dicts."""
+        # Filter the "nodePort" element from the port list.
+        config.filters = {"Service": [{"spec": [{"ports": ["nodePort"]}]}]}
+
+        expected: dict = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "metadata": {"name": "name", "namespace": "ns"},
+            "spec": {"type": "NodePort"},
+        }
+        manifest: dict = copy.deepcopy(expected)
+        manifest["spec"]["ports"] = [
+            {"name": "http", "port": 81, "nodePort": 1},
+            {"name": "http", "port": 82},
+            {"name": "http", "port": 83, "nodePort": 3},
+        ]
+        expected["spec"]["ports"] = [
+            {"name": "http", "port": 81},
+            {"name": "http", "port": 82},
+            {"name": "http", "port": 83},
+        ]
+        assert manio.strip_manifest(config, manifest) == expected
+
+    def test_strip_manifest_default_filters(self, config):
+        """Must fall back to default filters unless otherwise specified.
+
+        Here we expect the function to strip out the `metadata.uid` because it
+        is part of the default filters (see `square/resources/defaultconfig.yaml`).
+
+        """
+        manifest: Dict[str, Any] = {
+            "apiVersion": "v1",
+            "kind": "Service",
+            "something": "here",
+        }
+        expected = manifest.copy()
+
+        manifest["metadata"] = {"name": "name", "namespace": "ns", "uid": "some-uid"}
+        expected["metadata"] = {"name": "name", "namespace": "ns"}
+
+        # Must remove the `metadata.uid` field.
+        assert manio.strip_manifest(config, manifest)
